Backend/pyproject.toml
[project]
name = "elearning-backend"
version = "1.0.0"
description = "E-Learning Platform Backend API with RAG-powered Chatbot"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.3",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "llama-index>=0.11.20",
    "llama-index-llms-ollama>=0.3.4",
    "llama-index-embeddings-ollama>=0.3.1",
    "llama-index-vector-stores-chroma>=0.2.0",
    "chromadb>=0.5.20",
    "pypdf>=5.1.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",
]
